requirements.txt
pytest ==5.3.2
allure-pytest==2.7.0
mysql-connector-python==8.0.19
selenium